마이페이지 닉네임중복확인 로직

+++ src/controller/CheckmemAction.java
@@ -13,7 +13,7 @@ import itTalkDAO.MyPageDAO;
 import itTalkDO.Mb;
 
 public class CheckmemAction implements Action {
-
+	//비밀번호 변경 액션 & 회원탈퇴시 비밀번호 확인 액션
 	@Override
 	public ActionForward execute(HttpServletRequest req, HttpServletResponse res) throws ServletException, IOException {
 		MyPageDAO mypageDAO= new MyPageDAO();
@@ -26,21 +26,25 @@ public class CheckmemAction implements Action {
 		String paramPw =req.getParameter("pw");
 		System.out.println(paramPw);
 		
-		PrintWriter out=res.getWriter();
-		
 		req.setCharacterEncoding("UTF-8");
 		res.setContentType("text/html; charset=UTF-8");
+		
+		PrintWriter out=res.getWriter();
 		
 		if(paramPw.equals(myInfo.getMb_pw())) {
 			//본인인증 성공
 			System.out.println(myInfo.getMb_pw());
 			System.out.println(req.getParameter("c"));
 			
-			if(req.getParameter("c").equals("withdraw")) {
-				mypageDAO.deleteMember(1);
-				out.println("<script>alert('회원탈퇴 되었습니다.');</script>");
+			if(req.getParameter("c")!=null) {//회원탈퇴 액션일 때
+				
+				if(req.getParameter("c").equals("withdraw")) {
+					mypageDAO.deleteMember(1);
+					out.println("<script>alert('회원탈퇴 되었습니다.');</script>");
+				}
 			}
-			else {
+			
+			else {//비밀번호변경 액션일때
 				
 				out.println("<script>location.href='newpw2.jsp';</script>");
 			}
